Put routing error tests in own file
