+++ context/app/test_routes_main.py
@@ -4,7 +4,6 @@ from xml.etree.ElementTree import ParseError
 import json
 
 import pytest
-import requests
 
 from .main import create_app
 from .config import types
@@ -88,50 +87,6 @@ def test_200_json_page(client, path, mocker):
     json.loads(response.data.decode('utf8'))
 
 
-@pytest.mark.parametrize('path', [
-    '/no-page-here',
-    '/browse/no-such-type']
-    + [f'/browse/{t}/fake-uuid.fake' for t in types]
-)
-def test_404_html_page(client, path):
-    response = client.get(path)
-    assert response.status == '404 NOT FOUND'
-    assert_is_valid_html(response)
-    assert '404: Not Found' in response.data.decode('utf8')
-
-
-@pytest.fixture
-def client_not_logged_in():
-    app = create_app()
-    app.config['TESTING'] = True
-    with app.test_client() as client:
-        # No nexus_token!
-        yield client
-
-
-def test_403_html_page(client_not_logged_in):
-    response = client_not_logged_in.get('/browse/donor')
-    assert response.status == '403 FORBIDDEN'
-    assert_is_valid_html(response)
-    assert '403: Access Denied' in response.data.decode('utf8')
-
-
-def mock_get_400(path, **kwargs):
-    class MockResponse():
-        def __init__(self):
-            self.status_code = 400
-            self.text = 'Logger call requires this'
-        def raise_for_status(self):
-            raise requests.exceptions.HTTPError(response=self)
-    return MockResponse()
-
-
-def test_400_html_page(client, mocker):
-    mocker.patch('requests.get', side_effect=mock_get_400)
-    response = client.get('/browse/donor')
-    assert response.status == '400 BAD REQUEST'
-    assert_is_valid_html(response)
-    assert '400: Bad Request' in response.data.decode('utf8')
 
 
 def test_login(client):
